Padronizando variavel de distancia para float.

+++ source/include/CombatData.h
@@ -10,10 +10,10 @@ class CombatData
         CombatData(int nU1, int nU2);
         ~CombatData();
 
-        double getUnitDistance(const Unit* a, const Unit* b);
-//        double getShipDistance(const Ship* a, const Ship* b);
+        float getUnitDistance(const Unit* a, const Unit* b);
+//        float getShipDistance(const Ship* a, const Ship* b);
 
-        Unit* getNearestUnit(const Unit* from, const std::vector<Unit*>& to, double &outDist);
+        Unit* getNearestUnit(const Unit* from, const std::vector<Unit*>& to, float &outDist);
 //        Ship* getNearestShip(const Ship* from, const std::vector<Ship*>& to);
 
         //Unit* getWeakestUnit(const Unit* from, const std::vector<Unit*>& to);
@@ -27,7 +27,7 @@ class CombatData
         int miss, total;
 
         // [team0][team1]
-        double** distancesUnit;
+        float** distancesUnit;
         int nUnits1, nUnits2;
 };
